# FlashD/src/main/resources/static/swagger/index.yaml
swagger: '2.0'
info:
  description: jeefastis一个轻量级的Java快速开发平台
  version: 1.0.0
  title: jeefast快速开发平台
basePath: /jeefast
schemes:
  - http

#认证
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "token"
    in: "header"

#定义interfacedata
paths:
...
  /sys/user/list:
    get:
      tags:
        - USER
      summary: USERlist
      produces:
        - application/json
      parameters:
        - name: page
          description: 页码
          in: query
          type: integer
          required: true
        - name: limit
          description: 每页条数
          in: query
          type: integer
          required: true
        - name: sidx
          description: sort字段
          in: query
          type: string
        - name: order
          description: sort方式，如：asc、desc
          in: query
          type: string
        - name: username
          description: username
          in: query
          type: string
      responses:
        '200':
          description: return userlist
          schema:
            $ref: '#/definitions/SysUser'
...
  /sys/user/info/{userId}:
    get:
      tags:
        - USER
      summary: get userinfo
      produces:
        - application/json
      parameters:
        - name: userId
          description: userId
          in: path
          type: integer
          required: true
      responses:
        '200':
          description: return userinfo
          schema:
            $ref: '#/definitions/SysUser'
  /sys/user/password:
    post:
      tags:
        - USER
      summary: updatepassword
      produces:
        - application/json
      parameters:
        - name: password
          description: old password
          in: query
          type: string
          required: true
        - name: newPassword
          description: new password
          in: query
          type: string
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/user/save:
    post:
      tags:
        - USER
      summary: add user
      produces:
        - application/json
      parameters:
        - name: body
          description: manage员object
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysUserEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/user/update:
    post:
      tags:
        - USER
      summary: update user
      produces:
        - application/json
      parameters:
        - name: body
          description:  userobject
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysUserEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/user/delete:
    post:
      tags:
        - USER
      summary: delete user
      produces:
        - application/json
      parameters:
        - name: body
          description: userIdlist
          in: body
          type: array
          items:
            type: integer
            format: int64
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
definitions:
  R:
      type: object
      properties:
        code:
          description: state码 0：success  非0：失败
          type: integer
          format: int32
        msg:
          description: 失败原因
          type: string
  SysUser:
    type: object
    properties:
      userId:
        description: userId
        type: integer
        format: int64
      username:
        description: username
        type: string
      deptId:
        description: deptId
        type: integer
        format: int64
      deptName:
        description: Department Name
        type: string
      password:
        description: password
        type: string
      email:
        description: email
        type: string
      mobile:
        description: mobile
        type: string
      status:
        description: state  0：Disable   1：normal
        type: integer
        format: int32
      roleIdList:
        description: roleIDlist
        type: array
        items:
          type: integer
          format: int64
      createUserId:
        description: createUserID
        type: integer
        format: int64
      createTime:
        description: createTime
        type: string
        format: date-time
  SysUserEdit:
      type: object
      properties:
        userId:
          description: userId
          type: integer
          format: int64
        username:
          description: username
          type: string
        deptId:
          description: deptId
          type: integer
          format: int64
        password:
          description: password
          type: string
        email:
          description: email
          type: string
        mobile:
          description: mobile
          type: string
        status:
          description: state  0：Disable   1：normal
          type: integer
          format: int32
        roleIdList:
          description: roleIDlist
          type: array
          items:
            type: integer
            format: int32
...